On WP Single, always generate activation key
When 'bp_core_signup_send_activation_key' is filtered to false, an activation
key should still be generated, otherwise the newly created user will be unable
to activate her account. This changeset ensures that the activation key is
generated, even if a plugin is preventing the key from being sent via email.

Fixes #4218