Reinstates 1.5.x integration of bbPress 1.x forum installation and configuration
In anticipation for a full switchover to bbPress 2.x in BuddyPress 1.6, several
aspects of the forum configuration and setup process were modified. Since
the migration to bbPress 2.x is being put off to a later version of BuddyPress,
it's necessary to revert some of these changes back to their BP 1.5.x state.

This changeset also fixes a bug in the bb-config.php creation script, where
a salt containing the string '?>' could result in early termination of the
config file.

Fixes #4269